backend/requirements.txt
# ──────────────────────────────────────────────
# Core
# ──────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1

# ──────────────────────────────────────────────
# Azure SDKs
# ──────────────────────────────────────────────
azure-identity==1.19.0
azure-storage-file-datalake==12.18.0

# ──────────────────────────────────────────────
# Utilities
# ──────────────────────────────────────────────
python-dotenv==1.0.1
python-multipart==0.0.20
